+++ src/main/java/com/casatoronto/challenge/model/CasaTorontoModel.java
@@ -4,6 +4,8 @@ import org.springframework.data.mongodb.core.mapping.FieldType;
 import org.springframework.data.mongodb.core.mapping.MongoId;
 
 import com.casatoronto.challenge.model.audit.UserDateAudit;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 public abstract class CasaTorontoModel extends UserDateAudit {
 
@@ -18,5 +20,19 @@ public abstract class CasaTorontoModel extends UserDateAudit {
 
 	public void setId(String id) {
 		this.id = id;
+	}
+	
+	@Override
+	public String toString() {
+		ObjectMapper mapper = new ObjectMapper();
+		String jsonInString = "";
+
+		try {
+			jsonInString = mapper.writeValueAsString(this);
+		} catch (JsonProcessingException e) {
+			e.printStackTrace();
+		}
+
+		return jsonInString;
 	}
 }
