+++ src/main/java/com/casatoronto/challenge/payload/TravelRequest.java
@@ -1,6 +1,5 @@
 package com.casatoronto.challenge.payload;
 
-import java.io.Serializable;
 import java.time.LocalDate;
 
 import org.springframework.format.annotation.DateTimeFormat;
@@ -8,18 +7,16 @@ import org.springframework.lang.NonNull;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
- * Travel
+ * TravelRequest
  * 
- * class responsible for saving checkIn and checkOut information.
+ * class responsible for request checkIn and checkOut information.
  * 
  * @author ecalado@gmail.com
  *
  */
-public class TravelRequest implements Serializable {
+public class TravelRequest extends CasaTorontoRequest {
 
 	private static final long serialVersionUID = -5826595171668032872L;
 
@@ -44,19 +41,5 @@ public class TravelRequest implements Serializable {
 
 	public LocalDate getCheckOut() {
 		return checkOut;
-	}
-
-	@Override
-	public String toString() {
-		ObjectMapper mapper = new ObjectMapper();
-		String jsonInString = "";
-
-		try {
-			jsonInString = mapper.writeValueAsString(this);
-		} catch (JsonProcessingException e) {
-			e.printStackTrace();
-		}
-
-		return jsonInString;
 	}
 }
