--- src/main/java/com/casatoronto/challenge/security/CurrentUser.java
@@ -1,17 +1,0 @@
-package com.casatoronto.challenge.security;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-
-@Target({ ElementType.PARAMETER, ElementType.TYPE })
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@AuthenticationPrincipal
-public @interface CurrentUser {
-
-}
