--- src/main/java/com/casatoronto/challenge/security/JwtAuthenticationFilter.java
@@ -1,62 +1,0 @@
-package com.casatoronto.challenge.security;
-
-import java.io.IOException;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-	@Autowired
-	private JwtTokenProvider tokenProvider;
-
-	@Autowired
-	private CustomUserDetailsService customUserDetailsService;
-
-	private static final Logger logger = LoggerFactory.getLogger(JwtAuthenticationFilter.class);
-
-	@Override
-	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-			throws ServletException, IOException {
-		try {
-			String jwt = getJwtFromRequest(request);
-
-			if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
-				String userId = tokenProvider.getUserIdFromJWT(jwt);
-
-				UserDetails userDetails = customUserDetailsService.loadUserByUsername(userId);
-				UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(
-						userDetails, null, userDetails.getAuthorities());
-				authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-
-				SecurityContextHolder.getContext().setAuthentication(authentication);
-			}
-		} catch (Exception ex) {
-			logger.error("Could not set user authentication in security context", ex);
-			throw new UsernameNotFoundException("Could not set user authentication in security context", ex);
-		}
-
-		filterChain.doFilter(request, response);
-	}
-
-	private String getJwtFromRequest(HttpServletRequest request) {
-		String bearerToken = request.getHeader("Authorization");
-		if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-			return bearerToken.substring(7, bearerToken.length());
-		}
-		return null;
-	}
-}
