+++ src/main/java/com/casatoronto/challenge/security/config/JwtAuthenticationEntryPoint.java
@@ -1,0 +1,26 @@
+package com.casatoronto.challenge.security.config;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+	private static final Logger logger = LoggerFactory.getLogger(JwtAuthenticationEntryPoint.class);
+
+	@Override
+	public void commence(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse,
+			AuthenticationException e) throws IOException, ServletException {
+		logger.error("Responding with unauthorized error. Message - {}", e.getMessage());
+		httpServletResponse.sendError(HttpServletResponse.SC_UNAUTHORIZED, e.getMessage());
+	}
+}
