+++ src/main/java/com/casatoronto/challenge/service/RoleService.java
@@ -4,10 +4,9 @@ import java.util.Optional;
 
 import com.casatoronto.challenge.model.Role;
 import com.casatoronto.challenge.model.RoleName;
+import com.casatoronto.challenge.payload.RoleRequest;
 
-public interface RoleService {
+public interface RoleService extends CasaTorontoService<Role, RoleRequest, String> {
 
 	public Optional<Role> findByName(RoleName roleName);
-	
-	public Optional<Role> insert(Role role);
 }
