+++ src/main/java/com/casatoronto/challenge/service/UserService.java
@@ -3,17 +3,9 @@ package com.casatoronto.challenge.service;
 import java.util.Optional;
 
 import com.casatoronto.challenge.model.User;
-import com.casatoronto.challenge.payload.SignUpRequest;
+import com.casatoronto.challenge.payload.UserRequest;
 
-public interface UserService {
-
-	public Optional<User> findById(String id);
+public interface UserService extends CasaTorontoService<User, UserRequest, String> {
 
 	public Optional<User> findByUsernameOrEmail(String username, String email);
-
-	public User create(SignUpRequest signupRequest);
-
-	public User create(SignUpRequest signupRequest, User user);
-
-	public Optional<User> insert(SignUpRequest signupRequest);
 }
