+++ src/main/java/com/casatoronto/challenge/service/UserServiceImpl.java
@@ -3,7 +3,6 @@ package com.casatoronto.challenge.service;
 import java.util.Collections;
 import java.util.Optional;
 
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -14,7 +13,8 @@ import com.casatoronto.challenge.exception.UsernameInUseException;
 import com.casatoronto.challenge.model.Role;
 import com.casatoronto.challenge.model.RoleName;
 import com.casatoronto.challenge.model.User;
-import com.casatoronto.challenge.payload.SignUpRequest;
+import com.casatoronto.challenge.payload.UserRequest;
+import com.casatoronto.challenge.repository.CasaTorontoRepository;
 import com.casatoronto.challenge.repository.UserRepository;
 
 @Service
@@ -30,24 +30,18 @@ public class UserServiceImpl implements UserService {
 	private PasswordEncoder passwordEncoder;
 
 	@Override
-	public User create(SignUpRequest signupRequest) {
-		User user = new User();
-
-		this.create(signupRequest, user);
-
-		return user;
+	public User getInstanceOfEntityModel() {
+		return new User();
 	}
 
 	@Override
-	public User create(SignUpRequest signupRequest, User user) {
-		BeanUtils.copyProperties(signupRequest, user);
-
-		return user;
+	public CasaTorontoRepository<User, String> getRepository() {
+		return repository;
 	}
 
 	@Override
-	public Optional<User> insert(SignUpRequest signupRequest) {
-		User user = this.create(signupRequest);
+	public Optional<User> insert(UserRequest userRequest) {
+		User user = this.create(userRequest);
 
 		if (repository.existsByUsername(user.getUsername())) {
 			throw new UsernameInUseException();
@@ -68,13 +62,7 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public Optional<User> findById(String id) {
-		return repository.findById(id);
-	}
-
-	@Override
 	public Optional<User> findByUsernameOrEmail(String username, String email) {
 		return repository.findByUsernameOrEmail(username, email);
 	}
-
 }
